removed "switch trick"
In my experience it's not valuable, just annoying.  When adding a new
state or event, I always look at the diagram to check if I made all the
right connections.  Therefore, having this code to make me not forget
about it is not useful and updating it is just busywork.